Working Smallrye Config support
