+++ src/test/java/hu/rmegyesi/mpconfig/docgen/smallryeconfig/SmallryeConfigMappingAnnotationProcessorTest.java
@@ -1,4 +1,4 @@
-package hu.rmegyesi.mpconfig.docgen;
+package hu.rmegyesi.mpconfig.docgen.smallryeconfig;
 
 /*-
  * #%L
@@ -20,7 +20,8 @@ package hu.rmegyesi.mpconfig.docgen;
  * #L%
  */
 
-import hu.rmegyesi.mpconfig.test.mpconfig.AggregatedConfig;
+import hu.rmegyesi.mpconfig.docgen.data.ConfigPropertyDocElement;
+import hu.rmegyesi.mpconfig.test.smallryeconfig.NestedConfigMapping;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
@@ -31,16 +32,18 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-public class MPConfigAnnotationProcessorTest {
+import static hu.rmegyesi.mpconfig.docgen.TestUtils.getExpectedPropertyNames;
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(MPConfigAnnotationProcessor.class);
+public class SmallryeConfigMappingAnnotationProcessorTest {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(SmallryeConfigMappingAnnotationProcessorTest.class);
 
     @Test
     void testNestedConfig() {
-        MPConfigAnnotationProcessor processor = new MPConfigAnnotationProcessor();
-        Stream<ConfigPropertyDocElement> configs = processor.processClass(AggregatedConfig.class);
+        SmallryeConfigMappingAnnotationProcessor processor = new SmallryeConfigMappingAnnotationProcessor();
+        Stream<ConfigPropertyDocElement> configs = processor.processConfigMappingInterface(NestedConfigMapping.class);
 
-        Set<String> expectedNames = TestUtils.getExpectedPropertyNames(AggregatedConfig.class);
+        Set<String> expectedNames = getExpectedPropertyNames(NestedConfigMapping.class);
         Set<String> configNames = configs.map(ConfigPropertyDocElement::name).collect(Collectors.toSet());
 
         Set<String> unexpectedNames = new HashSet<>(configNames);
@@ -56,5 +59,4 @@ public class MPConfigAnnotationProcessorTest {
         Assertions.assertTrue(unexpectedNames.isEmpty());
         Assertions.assertTrue(missingNames.isEmpty());
     }
-
 }
